♻️ refactor: create cleanup.c
