+++ 2048.h
@@ -6,6 +6,7 @@
 #include <stdlib.h>
 #include <time.h>
 #include <unistd.h>
+#include <fcntl.h>
 
 #include <ncurses.h>
 
@@ -71,42 +72,44 @@ typedef enum e_result
 	SCORE_LIST_SIZE_ERROR,
 }	t_result;
 
+//init
+t_result	init_score(t_board *board);
+void		init_board(t_board *board);
 
 // ncusrses
 t_result	init_ncurses(t_board *board);
-t_result	init_score(t_board *board);
-void	cleanup_ncurses(t_board *board);
-void	update_board(t_board *board);
-void	update_score(t_board *board);
+void		update_board(t_board *board);
+void		update_score(t_board *board);
+
 // main game logic
-void	init_board(t_board *board);
-void	game_loop(t_board *board);
-void	check_game_over(t_board *board);
-// number spawning
-void	spawn_number(t_board *board);
-// move and merge handler
-void	exec_move(t_board *board, t_direction dir);
+void		game_loop(t_board *board);
+void		check_game_over(t_board *board);
+void		spawn_number(t_board *board);
+void		exec_move(t_board *board, t_direction dir);
+
 // logic helpers for move & merge
-void	traverse_board(
-			t_board *board,
-			t_direction dir,
+void		traverse_board(
+	t_board *board,
+	t_direction dir,
 			void (*move)(t_board *, int, int),
 			void (*merge)(t_board *, int, int));
-void	reset_merged(t_board *board);
+void		reset_merged(t_board *board);
+
 // move and merge directions
-void	move_left(t_board *board, int x, int y);
-void	move_right(t_board *board, int x, int y);
-void	move_up(t_board *board, int x, int y);
-void	move_down(t_board *board, int x, int y);
-void	merge_left(t_board *board, int x, int y);
-void	merge_right(t_board *board, int x, int y);
-void	merge_up(t_board *board, int x, int y);
-void	merge_down(t_board *board, int x, int y);
+void		move_left(t_board *board, int x, int y);
+void		move_right(t_board *board, int x, int y);
+void		move_up(t_board *board, int x, int y);
+void		move_down(t_board *board, int x, int y);
+void		merge_left(t_board *board, int x, int y);
+void		merge_right(t_board *board, int x, int y);
+void		merge_up(t_board *board, int x, int y);
+void		merge_down(t_board *board, int x, int y);
 
-// debug functions
-void	debug_print(t_board *board);
-void	debug_move(t_board *board);
-
-
+// cleanup
 t_result	write_score_to_file(t_board *board);
 t_result	print_error(t_result result);
+void		cleanup_ncurses(t_board *board);
+
+// debug functions
+void		debug_print(t_board *board);
+void		debug_move(t_board *board);
